+++ packages/datasource-rpc/src/index.ts
@@ -1,93 +1,145 @@
 import { DataSourceFactory, Logger } from '@forestadmin/datasource-toolkit';
-import crypto from 'crypto';
-import { EventSource } from 'eventsource';
 import superagent from 'superagent';
 
 import RpcDataSource from './datasource';
-import { RpcDataSourceOptions, RpcSchema } from './types';
-import { appendHeaders, getAuthoriztionHeaders } from './utils';
+import { IntrospectionSchema, RpcDataSourceOptions, RpcSchema } from './types';
+import { appendHeaders, cameliseKeys, toPascalCase } from './utils';
 
 export { reconciliateRpc } from './plugins';
 
-const sseMap = new Map<string, EventSource>();
+/** polling interval in second */
+const DEFAULT_POLLING_INTERVAL = 600;
+const MIN_POLLING_INTERVAL = 1;
+const MAX_POLLING_INTERVAL = 3600;
 
-export async function getintrospection(
+function parseIntrospection(introSchema: IntrospectionSchema): RpcSchema {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  const { collections, charts, rpc_relations, native_query_connections, etag } = introSchema;
+
+  const parsedCollections = collections.map(collection => {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    const { actions, fields, aggregation_capabilities, ...rest } = collection;
+    const parsedActions = Object.entries(actions).reduce((pActions, [name, schema]) => {
+      pActions[name] = cameliseKeys(schema);
+
+      return pActions;
+    }, {});
+
+    const parsedFields = Object.entries(fields).reduce((pFields, [name, schema]) => {
+      pFields[name] = cameliseKeys(schema);
+      pFields[name].filterOperators = new Set(pFields[name].filterOperators.map(toPascalCase));
+
+      return pFields;
+    }, {});
+
+    return {
+      ...rest,
+      aggregationCapabilities: {
+        supportedDateOperations: new Set(aggregation_capabilities.supported_date_operations),
+        supportGroups: aggregation_capabilities.support_groups,
+      },
+      actions: parsedActions,
+      fields: parsedFields,
+    };
+  });
+
+  const parsedRelations = Object.entries(rpc_relations).reduce(
+    (rpcRelations, [collectionName, collectionRelations]) => {
+      rpcRelations[collectionName] = Object.entries(collectionRelations).reduce(
+        (relations, [name, schema]) => {
+          relations[name] = cameliseKeys(schema);
+
+          return relations;
+        },
+        {},
+      );
+
+      return rpcRelations;
+    },
+    {},
+  );
+
+  return {
+    collections: parsedCollections,
+    charts,
+    rpcRelations: parsedRelations,
+    nativeQueryConnections: native_query_connections,
+    etag,
+  };
+}
+
+async function getIntrospection(
   logger: Logger,
   uri: string,
   authSecret: string,
+  etag?: string,
 ): Promise<RpcSchema> {
   logger('Info', `Getting schema from Rpc agent on ${uri}.`);
 
   const introRq = superagent.get(`${uri}/forest/rpc-schema`);
   appendHeaders(introRq, authSecret);
+  if (etag) introRq.set('if-none-match', etag);
 
   const introResp = await introRq.send();
 
-  return introResp.body;
+  return parseIntrospection(introResp.body);
 }
 
-function getHash(schema: RpcSchema) {
-  return crypto.createHash('sha1').update(JSON.stringify(schema)).digest('hex');
-}
-
-function runRecon(
+function startPolling(
   logger: Logger,
-  uri: string,
-  authSecret: string,
-  originalHash: string,
-  restartAgent: () => Promise<void>,
+  options: { uri: string; authSecret: string; etag: string; pollingInterval?: number },
+  onChange: () => Promise<void>,
 ) {
-  if (!sseMap.has(uri)) {
-    const es = new EventSource(`${uri}/forest/sse`, {
-      fetch: (input, init) => {
-        return fetch(input, {
-          ...init,
-          headers: {
-            ...init.headers,
-            ...getAuthoriztionHeaders(authSecret),
-          },
-        });
-      },
-    });
+  const { uri, authSecret, etag, pollingInterval } = options;
+  logger(
+    'Debug',
+    `Starting polling every ${pollingInterval} seconds for Rpc agent schema changes on ${uri}.`,
+  );
 
-    let reconnecting = false;
+  const inter = setInterval(async () => {
+    try {
+      const intro = await getIntrospection(logger, uri, authSecret, etag);
 
-    es.onerror = error => {
-      reconnecting = true;
-      logger('Debug', `SSE (${uri}) error: ${error.message}`);
-    };
-
-    es.onopen = async () => {
-      if (reconnecting) {
-        logger('Info', `Reconnecting with RPC agent on ${uri}.`);
-        const newIntrospection = await getintrospection(logger, uri, authSecret);
-        const newHash = getHash(newIntrospection);
-
-        if (originalHash !== newHash) {
-          logger('Info', `Schema of RPC agent on ${uri} change: restarting.`);
-          await restartAgent();
-        }
+      if (etag !== intro?.etag) {
+        logger('Info', `Schema change detected on Rpc agent ${uri}. Restarting agent.`);
+        clearInterval(inter);
+        onChange();
       }
-    };
-
-    sseMap.set(uri, es);
-  }
+    } catch (error) {
+      if (error.status === 304) {
+        logger('Debug', `No schema change detected on Rpc agent ${uri}.`);
+      } else if (!error.status) {
+        logger('Error', `Error while polling Rpc agent ${uri} for schema changes: Unreachable.`);
+      } else {
+        logger(
+          'Error',
+          `Error while polling Rpc agent ${uri} for schema changes: ${error.message}.`,
+        );
+      }
+    }
+  }, pollingInterval * 1000);
 }
 
 export function createRpcDataSource(options: RpcDataSourceOptions): DataSourceFactory {
   return async (logger: Logger, restartAgent: () => Promise<void>) => {
     const { authSecret, uri } = options;
-    let { introspection } = options;
+    const { introspection } = options;
+    let schema: RpcSchema;
 
-    if (!introspection) {
-      introspection = await getintrospection(logger, uri, authSecret);
+    try {
+      schema = await getIntrospection(logger, uri, authSecret);
+    } catch (error) {
+      if (!introspection) throw error;
+
+      schema = parseIntrospection(introspection);
     }
 
-    if (options.disableSSE) {
-      const originalHash = getHash(introspection);
-      runRecon(logger, uri, authSecret, originalHash, restartAgent);
-    }
+    const pollingInterval = Math.min(
+      Math.max(options.pollingInterval ?? DEFAULT_POLLING_INTERVAL, MIN_POLLING_INTERVAL),
+      MAX_POLLING_INTERVAL,
+    );
+    startPolling(logger, { uri, authSecret, etag: schema.etag, pollingInterval }, restartAgent);
 
-    return new RpcDataSource(logger, options, introspection);
+    return new RpcDataSource(logger, options, schema);
   };
 }